requirements.txt
fastapi
uvicorn[standard]
facenet-pytorch
torch
torchvision
opencv-python
Pillow
numpy
timm
einops
